// apps/admin/types/portfolio.ts
export enum ProgrammingRole {
  FullStack = "FullStack",
  Frontend = "Frontend",
  Backend = "Backend",
  Mobile = "Mobile",
}

export enum ProgrammingLevel {
  Beginner = "Beginner",
  Intermediate = "Intermediate",
  Advanced = "Advanced",
  Expert = "Expert",
}

export enum JobType {
  Full_Time = "Full_Time",
  Part_Time = "Part_Time",
  Contract = "Contract",
  Freelance = "Freelance",
}
// ...
export interface Contact {
  id?: string;
  externalId?: string;
  name: string;
  icon: string;
  link: string;
}

export interface Feature {
  id?: string;
  externalId?: string;
  title: string;
  description: string;
  techStacks: TechStack[];
}

export interface Project {
  id?: string;
  externalId?: string;
  title: string;
  description: string;
  link: string;
  features: Feature[];
}

export interface Experience {
  id?: string;
  externalId?: string;
  company_name: string;
  company_description: string;
  start_date: string;
  end_date: string | null;
  role: ProgrammingRole;
  job_type: JobType;
  contacts: Contact[];
}